module front end more dynamic in run panel
