+++ app/src/components/ModuleLiveStatusCard/TempDeckCard.js
@@ -1,6 +1,7 @@
 // @flow
 import * as React from 'react'
 import { LabeledValue } from '@opentrons/components'
+import { getModuleDisplayName } from '@opentrons/shared-data'
 
 import type { TempDeckModule } from '../../robot-api'
 import StatusCard from './StatusCard'
@@ -12,15 +13,18 @@ type Props = {|
 |}
 
 const TempDeckCard = ({ module }: Props) => (
-  <StatusCard title={module.displayName}>
+  <StatusCard title={getModuleDisplayName(module.name)}>
     <CardContentRow>
-      <StatusItem status={status} />
+      <StatusItem status={module.status} />
     </CardContentRow>
     <CardContentRow>
-      <LabeledValue label="Current Temp" value={`${module.currentTemp} °C`} />
+      <LabeledValue
+        label="Current Temp"
+        value={`${module.data.currentTemp} °C`}
+      />
       <LabeledValue
         label="Target Temp"
-        value={module.targetTemp ? `${module.targetTemp} °C` : 'None'}
+        value={module.data.targetTemp ? `${module.data.targetTemp} °C` : 'None'}
       />
     </CardContentRow>
   </StatusCard>
