+++ app/src/components/ModuleLiveStatusCard/index.js
@@ -2,14 +2,10 @@
 import * as React from 'react'
 import { connect } from 'react-redux'
 
-import { getModuleDisplayName } from '@opentrons/shared-data'
 import { getConnectedRobot } from '../../discovery'
 import { fetchModules, getModulesState } from '../../robot-api'
 
-import { LabeledValue, IntervalWrapper } from '@opentrons/components'
-import StatusCard from './StatusCard'
-import CardContentRow from './CardContentRow'
-import StatusItem from './StatusItem'
+import { IntervalWrapper } from '@opentrons/components'
 
 import type { State, Dispatch } from '../../types'
 import type { TempDeckModule } from '../../robot-api'
@@ -36,44 +32,37 @@ type Props = {|
 |}
 
 const ModuleLiveStatusCards = (props: Props) => {
-    const { liveStatusModules } = this.props
-    if (liveStatusModules.length === 0) return null
+  const { liveStatusModules, fetchModules } = props
+  if (liveStatusModules.length === 0) return null
 
-    const { name, status, data } = tempDeck
-    const { currentTemp, targetTemp } = data
-    const displayName = getModuleDisplayName(name)
-
-    return (
-      <IntervalWrapper
-        refresh={fetchModules}
-        interval={POLL_TEMPDECK_INTERVAL_MS}
-      >
-        {liveStatusModules.map(module => {
-          switch (module.name) {
-            case 'tempdeck':
-              return <TempDeckCard module={module} />
-            case 'thermocycler':
-              return <ThermocyclerCard module={module} />
-            case 'magdeck':
-            default:
-              return null
-          }
-        })}
-      </IntervalWrapper>
-    )
-  }
+  return (
+    <IntervalWrapper
+      refresh={fetchModules}
+      interval={POLL_TEMPDECK_INTERVAL_MS}
+    >
+      {liveStatusModules.map(module => {
+        switch (module.name) {
+          case 'tempdeck':
+            return <TempDeckCard module={module} />
+          case 'thermocycler':
+            return <ThermocyclerCard module={module} />
+          case 'magdeck':
+          default:
+            return null
+        }
+      })}
+    </IntervalWrapper>
+  )
 }
 
 function mapStateToProps(state: State): SP {
   const _robot = getConnectedRobot(state)
   const modules = _robot ? getModulesState(state, _robot.name) : []
 
-  // TOD0 (ka 2018-7-25): Only supporting 1 temp deck at a time at launch
   const liveStatusModules = modules.filter(m =>
     LIVE_STATUS_MODULES.includes(m.name)
   )
 
-  // bogus ternary to satisfy flow
   return {
     _robot,
     liveStatusModules,
@@ -88,7 +77,7 @@ function mapDispatchToProps(dispatch: Dispatch): DP {
 
 function mergeProps(stateProps: SP, dispatchProps: DP): Props {
   const { _fetchModules } = dispatchProps
-  const { _robot, liveStatusModules} = stateProps
+  const { _robot, liveStatusModules } = stateProps
 
   return {
     liveStatusModules,
